Refactor Linked Node serializer to base serializers
