+++ api/collections/serializers.py
@@ -5,9 +5,9 @@ from framework.exceptions import PermissionsError
 from website.exceptions import NodeStateError
 
 from website.models import Node
-from api.base.serializers import LinksField, RelationshipField, JSONAPIRelationshipSerializer
-from api.base.serializers import JSONAPISerializer, IDField, TypeField, relationship_diff
-from api.base.exceptions import InvalidModelValueError, RelationshipPostMakesNoChanges
+from api.base.serializers import LinksField, RelationshipField
+from api.base.serializers import JSONAPISerializer, IDField, TypeField
+from api.base.exceptions import InvalidModelValueError
 from api.base.utils import absolute_reverse, get_user_auth
 from api.nodes.serializers import NodeLinksSerializer
 
@@ -107,74 +107,3 @@ class CollectionNodeLinkSerializer(NodeLinksSerializer):
         )
 
 
-class LinkedNode(JSONAPIRelationshipSerializer):
-    id = ser.CharField(source='node._id', required=False, allow_null=True)
-    class Meta:
-        type_ = 'linked_nodes'
-
-
-class CollectionLinkedNodesRelationshipSerializer(ser.Serializer):
-
-    data = ser.ListField(child=LinkedNode())
-    links = LinksField({'self': 'get_self_url',
-                        'html': 'get_related_url'})
-
-    def get_self_url(self, obj):
-        return obj['self'].linked_nodes_self_url
-
-    def get_related_url(self, obj):
-        return obj['self'].linked_nodes_related_url
-
-    class Meta:
-        type_ = 'linked_nodes'
-
-    def get_pointers_to_add_remove(self, pointers, new_pointers):
-        diff = relationship_diff(
-            current_items={pointer.node._id: pointer for pointer in pointers},
-            new_items={val['node']['_id']: val for val in new_pointers}
-        )
-
-        nodes_to_add = []
-        for node_id in diff['add']:
-            node = Node.load(node_id)
-            if not node:
-                raise exceptions.NotFound(detail='Node with id "{}" was not found'.format(node_id))
-            nodes_to_add.append(node)
-
-        return nodes_to_add, diff['remove'].values()
-
-    def make_instance_obj(self, obj):
-        # Convenience method to format instance based on view's get_object
-        return {'data': [
-            pointer for pointer in
-            obj.nodes_pointer
-            if not pointer.node.is_deleted and not pointer.node.is_collection
-        ], 'self': obj}
-
-    def update(self, instance, validated_data):
-        collection = instance['self']
-        auth = get_user_auth(self.context['request'])
-
-        add, remove = self.get_pointers_to_add_remove(pointers=instance['data'], new_pointers=validated_data['data'])
-
-        for pointer in remove:
-            collection.rm_pointer(pointer, auth)
-        for node in add:
-            collection.add_pointer(node, auth)
-
-        return self.make_instance_obj(collection)
-
-    def create(self, validated_data):
-        instance = self.context['view'].get_object()
-        auth = get_user_auth(self.context['request'])
-        collection = instance['self']
-
-        add, remove = self.get_pointers_to_add_remove(pointers=instance['data'], new_pointers=validated_data['data'])
-
-        if not len(add):
-            raise RelationshipPostMakesNoChanges
-
-        for node in add:
-            collection.add_pointer(node, auth)
-
-        return self.make_instance_obj(collection)
